--- .github/workflows/embed-binaries.yaml
@@ -1,19 +1,0 @@
-name: Embed Binaries
-
-on:
-  schedule:
-    - cron: "59 23 * * *"
-  workflow_dispatch:
-
-permissions:
-  issues: write
-
-jobs:
-  embed-binaries:
-    uses: devantler-tech/workflows/.github/workflows/dotnet-embed-binaries.yaml@main
-    secrets: inherit
-    with:
-      BINARY_NAME: kustomize
-      DOTNET_PROJECT_NAME: Devantler.KustomizeCLI
-      SCRIPT_PATH: scripts/get-kustomize-binary.sh
-      VERSION_ARGS: version
