+++ DevantlerTech.KustomizeCLI.slnx
@@ -1,0 +1,13 @@
+<Solution>
+  <Configurations>
+    <Platform Name="Any CPU" />
+    <Platform Name="x64" />
+    <Platform Name="x86" />
+  </Configurations>
+  <Folder Name="/src/">
+    <Project Path="src/DevantlerTech.KustomizeCLI/DevantlerTech.KustomizeCLI.csproj" />
+  </Folder>
+  <Folder Name="/tests/">
+    <Project Path="tests/DevantlerTech.KustomizeCLI.Tests/DevantlerTech.KustomizeCLI.Tests.csproj" />
+  </Folder>
+</Solution>
